refactored the code, created an endpoint for fetching channel members instead of a route
