--- server/src/routes/channelMembers.route.ts
@@ -1,8 +1,0 @@
-import express from "express";
-import { fetchChannelMemberHandler } from "../controllers/channelMembers.controller";
-
-const router = express.Router();
-
-router.get("/:channel_id", fetchChannelMemberHandler);
-
-export default router;
